Bug fixing
