+++ app/src/main/java/com/bdp/onroad/AuthActivity.java
@@ -1,0 +1,61 @@
+package com.bdp.onroad;
+
+import android.content.Intent;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.widget.Toast;
+
+import com.firebase.ui.auth.AuthUI;
+import com.firebase.ui.auth.IdpResponse;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class AuthActivity extends AppCompatActivity {
+
+    List<AuthUI.IdpConfig> providers;
+    private final static int MyReqCode=2;
+    FirebaseAuth firebaseAuth;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_auth2);
+
+        providers = Arrays.asList(
+                new AuthUI.IdpConfig.EmailBuilder().build(),
+                new AuthUI.IdpConfig.GoogleBuilder().build(),
+                new AuthUI.IdpConfig.FacebookBuilder().build()
+        );
+        showSignInOptions();
+    }
+    public void showSignInOptions(){
+
+        startActivityForResult(
+
+                AuthUI.getInstance().createSignInIntentBuilder()
+                        .setAvailableProviders(providers)
+                        .setTheme(R.style.MyTheme)
+                        .build(),MyReqCode
+        );
+    }
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(requestCode == MyReqCode){
+            IdpResponse response = IdpResponse.fromResultIntent(data);
+            if(resultCode == RESULT_OK){
+                FirebaseUser user = firebaseAuth.getInstance().getCurrentUser();
+                Toast.makeText(AuthActivity.this, "Hi " + user.getDisplayName(), Toast.LENGTH_SHORT).show();
+                startActivity(new Intent(AuthActivity.this, UserTypeActivity.class));
+            }
+            else{
+                Toast.makeText(AuthActivity.this, ""+response.getError(), Toast.LENGTH_SHORT).show();
+            }
+        }
+    }
+}
